Move get_sensor_params into City
